Move plugin to vagrant-solidus gem, add solidus-box command

+++ lib/vagrant-solidus/site_helpers.rb
@@ -1,10 +1,11 @@
 module VagrantPlugins
-  module CommandSite
+  module Solidus
     module SiteHelpers
       BASE_PORT = 8081
       BASE_LIVERELOAD_PORT = 35730
       SITE_TEMPLATE_GIT_URL = "https://github.com/solidusjs/solidus-site-template.git"
       SITE_STATUS_WATCHER_POLLING_FREQUENCY = 1
+      PROVISION_ID = 20140502
 
       #########################################################################
       # System Calls
@@ -66,7 +67,7 @@ module VagrantPlugins
       def load_site
         @site_host_path           = File.join(ROOT_HOST_PATH, @site_name)
         @site_guest_path          = File.join(ROOT_GUEST_PATH, @site_name)
-        @site_log_file_path       = ".solidus-devbox/log/#{@site_name}.log"
+        @site_log_file_path       = ".vagrant-solidus/log/#{@site_name}.log"
         @site_log_file_guest_path = File.join(ROOT_GUEST_PATH, @site_log_file_path)
 
         if config = sites[@site_name]
@@ -126,11 +127,15 @@ module VagrantPlugins
         if File.exists?(File.join(@site_host_path, 'Gemfile'))
           return unless guest_exec(:log_on_error, "cd #{@site_guest_path} && bundle install")
         else
+          # Until all sites use bundler...
           return unless guest_exec(:log_on_error, "gem install sass")
         end
 
-        # Node packages
-        return unless guest_exec(:log_on_error, "npm install bower -g")
+        # Bower packages
+        if File.exists?(File.join(@site_host_path, 'bower.json'))
+          return unless guest_exec(:log_on_error, "npm install bower -g")
+          return unless guest_exec(:log_on_error, "cd #{@site_guest_path} && bower install")
+        end
 
         return true
       end
@@ -256,7 +261,7 @@ module VagrantPlugins
         opts.on("-f", "--fast", "Fast mode. Don't install the site first.") do
           @fast = true
         end
-        opts.on("-d", "--deaf", "Don't listen to host file events and forward them to the guest (events will be much slower).") do
+        opts.on("-d", "--deaf", "Don't automatically launch the `watch` command in background (file events will be much slower).") do
           @deaf = true
         end
       end
@@ -278,6 +283,18 @@ module VagrantPlugins
           @site_template_host_path  = File.join(ROOT_HOST_PATH, path)
           @site_template_guest_path = File.join(ROOT_GUEST_PATH, path)
         end
+      end
+
+      #########################################################################
+      # Provision
+      #########################################################################
+
+      def provisioned?
+        guest_exec(nil, "echo #{PROVISION_ID} | diff - ~/.vagrant-solidus/provision")
+      end
+
+      def provisioned!
+        "mkdir -p ~/.vagrant-solidus && echo #{PROVISION_ID} > ~/.vagrant-solidus/provision"
       end
 
       #########################################################################
